addvendorRouterUpdateDB
