+++ sqls/_connect_db.js
@@ -1,0 +1,15 @@
+const mysql = require('mysql');
+const bluebird = require('bluebird');
+const { database, username, password, host } = require(__dirname + '/../config/dbConfig');
+
+const db = mysql.createConnection({
+    host: host,
+    user: username,
+    password: password,
+    database: database
+});
+
+db.connect();
+bluebird.promisifyAll(db);
+
+module.exports = db;
